move test to linked apps test file
